requirements.txt
fastapi
uvicorn
python-dotenv
google-cloud-aiplatform
google-adk
google-cloud-modelarmor
requests
pydantic
cloudpickle
